pyproject.toml
[project]
name = "dress-composition-backend"
version = "1.0.0"
description = "의상 합성 백엔드 API 서버"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "transformers==4.35.2",
    "torch==2.1.1+cu118",
    "torchvision==0.16.1+cu118",
    "pillow>=10.0.0",
    "python-multipart==0.0.6",
    "jinja2==3.1.2",
    "numpy>=2,<2.3",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pymysql>=1.1.0",
    "sqlalchemy>=2.0.23",
    "cryptography>=41.0.0",
    "boto3>=1.34.0",
    "google-genai>=0.2.0",
    "mmpose>=0.31.0",
    "mmcv>=2.0.0",
    "mmengine>=0.10.0",
    "diffusers>=0.21.0",
    "basicsr>=1.4.0",
    "realesrgan>=0.3.0",
    "opencv-python>=4.8.0",
    "scipy>=1.11.0",
    "protobuf>=4.25.3,<5",
    "torchaudio==2.1.1+cu118",
    "mediapipe==0.10.21",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv.extra-build-dependencies]
xtcocotools = ["numpy"]
chumpy = ["pip", "setuptools", "wheel"]

[tool.setuptools.packages.find]
include = [
    "core",
    "list",
    "model",
    "myenv",
    "config",
    "static",
    "routers",
    "schemas",
    "weights",
    "services",
    "templates",
    "migrations",
    "body_analysis_test",
    "image_enhancement_server"
]
